busybox: tidy up variant
The variant was created by duplicating the juno-oe variant.

Now that it's working, this patch removes duplication from the juno-oe
and juno-busybox scripts.

Due to the nature that the variant scripts are processed, the juno-oe
and juno-android scripts needs to include the juno-busybox script,
rather than the other way round, otherwise the packaging step fails.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>