Preserve origin DTB when adding chosen node
The code was over-writing the original DTB file when adding the chosen
node.  However, this would be a problem when trying out different
ramdisks or doing other things that would change the size of the
ramdisk, eg, building a variant that used an initrd followed by one that
uses a uInitrd.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>