arm-tf: Allow to configure build type from command line
Instead of hard-coding the build type (debug/release) in the variant
file, allow to configure it from the command line. If it is not
specified on the command line then the variant file sets its default
value.

Also rename 'ARMTF_DEBUG_ENABLED' into 'ARM_TF_DEBUG_ENABLED' to
align it with the other Trusted Firmware environment variables.

Signed-off-by: Sandrine Bailleux <sandrine.bailleux@arm.com>