ARMLT: EAS test kernel support for all variants.
This patch is for the ARMLT tree only.

If the linux kernel config fragments for EAS are present, use them.
Also let the user provide BL_SUPPORT=EAS or BL_SUPPORT=big-LITTLE-MP
variable to the build scripts if desired.

Will warn and try to provide help if your chosen config does not exist
in the kernel source.

Change-Id: I9a850d03f00741219fb1b45e2e743b6098ec4669
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>