Move BL_SUPPORT config to end of list
The BL_SUPPORT variable is set to the config fragment the build uses for
the kernel.

Move this variable to the end of the list so it over-rides all other
fragments.

Currently, Android or OE choices could revert configs we request in
EAS.conf of big-LITTLE-MP.conf and that should not happen.

We also specifically want it to sit after the platform config.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>