Fix cert_create params
The following ARM-TF commit changes the params needed for cert_create:

96103d5  2016-01-22  cert_create: add non-volatile counter support [Juan Castillo]

New params are needed to continue to be able to use cert_create as
before.  However, it has been pointed out that we shouldn't use
cert_create directly, but instead, we should invoke it via the makefile
which deals with the default params.

Change-Id: I6f66b3faa5386b3c91cb60950efd25dbf1b5af67
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>