Change to build-scripts arguments and structure
If there is only one filesystem/flavour then the -f and -t options
become optional, and omitting them will be the same as specifying the
only available filesystem/flavour with those flags. Also changing so
that if there is only one platform/filesystem/flavour, then
"./build-$COMP.sh $CMD" will work with no options (where $COMP is a
component, or 'all' and $CMD is clean/build/package).

Splitting the public 'platform' into the real platforms. The folder
platforms/public no longer exists and in it's place, there are separate
juno, tc2 and fvp.

Changing the output directory to be "output/platform" for platforms with
only one flavour and "output/platform/flavour" for platforms with
multiple flavours.

Previously, there was a huge amount of find commands to handle
platform/filesystem/flavour detection over these 3 files. They all
needed to be modified to ignore hidden objects, so it was a good time to
ensure that they were abstracted.

Change-Id: Id6054966651d39251665a960047217466c06fb88
Signed-off-by: Dean Birch <dean.birch@arm.com>