build-busybox: use CROSS_COMPILE_32 for arm builds
When ARCH=arm, build-busybox was still using the default CROSS_COMPILE
to build, meaning that on aarch64 platforms, it was built with the wrong
compiler.

Whilst this worked on aarch64 platforms, packaging for an aarch32
platform without rebuilding busybox left the aarch32 system unbootable.

Conversely, an aarch64 system will boot with an ARCH=arm busybox built
using the aarch32 compiler.

Change-Id: I74269d8bd2662165155681ecf20588412bfb7b07
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>