Create common.fvp to share between fvp and fvp32
THe workspace sync script provided by PEG deletes the platforms not used
by the user.

Therefore, we need to keep the file included from fvp32 or it will
break.

To do this, make a new common file, platforms/common/common.fvp and
include it from both the fvp and fvp32 platform configs.

Change-Id: Ie8f951a19733c19bfed53f6e8860566884f683ce
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>