build-arm-tf.sh: use SPD=optee only on Juno
Building wth SPD=opteed tells ARM-TF that there is a BL32. However, we
weren't supplying a BL32 for FVP, only for Juno: OPTEE is disabled on
FVP.

Sine a recent ARM-TF commit, FVP hangs on boot when expecting to use
BL32.

To fix this problem, only set SPD=opteed for Juno builds.

Change-Id: I857fcb30d5a3c9c1bc0a592712ca0dd5dac8bcb3
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>