filesystems: allow build without any specific filesystem specified
For some of the test cases such as distro-boot, it is not mandatory
to build any of the supported filesystems. For such test cases, a
new 'none' filesystem is introduced. The configuration file for the
'none' filesystem is an empty file. This approach is the least
intrusive approach to avoid building filesystems for test cases
that do not need any specific filesystem to be built.

Change-Id: I8368a01cfdab33ec1bbe8da3568a5916a51d7627
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>