sgi/build-test-kvm: add build support for validating kvm on sgi platforms
Add support for building the fip and disk images for validating kvm
functionality on Arm's SGI platforms. Except for grub, all the other
components gets built as part of the build process.

The build requires a prebuilt fedora disk image. The compiled linux
kernel image is loaded into this prebuilt disk image during the
package stage. The kvmtool required to launch the virtual machines
is expected to present in the fedora disk image.

Change-Id: Ia0d446e549922680529f708ad13932b034a053b7
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>