configs: rename 'platforms' directory as 'configs'
As build-scripts scales up to support platforms specfic tests, there is
a need to create a platform specific directory to hold the platform
specfic test build scripts. One of the ways to organize the platform
specfic test build scripts could be as below

	build-scripts
		|
		|-- build-<component1>.sh
		|-- build-<component2>.sh
		|-- .
		|-- .
		|-- build-<componentX>.sh
		|-- build-test-TEST1.sh
		|-- build-test-TEST2.sh
		|-- .
		|-- .
		|-- build-test-TESTX.sh
		|
		|-- configs
		|	|
		|	|-- <platform1>
		|	|-- <platform2>
		|	|-- .
		|	|-- .
		|	|-- <platformX>
		|
		|-- <platform1>
		|	|
		|	|-- <build-test-TEST1>.sh
		|	|-- <build-test-TEST2>.sh
		|	|-- .
		|	|-- .
		|	|-- <build-test-TESTX>.sh
		|
		|-- <platform2>
		|	|
		|	|-- <build-test-TEST1>.sh
		|	|-- <build-test-TEST2>.sh
		|	|-- .
		|	|-- .
		|	|-- <build-test-TESTX>.sh
		|
		|-- <platformX>
			|
			|-- <build-test-TEST1>.sh
			|-- <build-test-TEST2>.sh
			|-- .
			|-- .
			|-- <build-test-TESTX>.sh

The existing 'platforms' directory is in fact a bunch of build config
files. So the 'platforms' directory can be renamed as 'configs'
directory to allow adding platform specific directory under the root
of build-scripts directory. This would help to avoid any possible
incorrect interpretation of the purpose of the configs directory
and the platform specific test build scripts directory.

Change-Id: I3ca0576cfdbc7be88d2812fcda841c19c95ce8ca
Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>