build-uefi: allow platforms to override name of the built image
In preparation to add the build of uefi StandaloneMM binary, allow
platforms to override the name of the built image. When building
StandaloneMM along with uefi, there are two build targets, one for
non-secure uefi and one for StandaloneMM. So instead of fixing
the built image name to 'uefi.bin', allow platform config file
to override the name of the built image.

Change-Id: If6b84ae5801b570a52bf6afa78edec76f0a9f0fa
Signed-off-by: Sughosh Ganu <sughosh.ganu@arm.com>