build-target-bins: fix invalid indirect expansion
Using build-target-bins on Ubuntu 20.04 results in the following error:
"uboot_out: invalid indirect expansion"
Ubuntu 20.04 LTS uses bash version 5.0 which doesn't expand an indirect
expansion if the variable is not defined, which caused the above error.
Fix this by checking if the uboot_out variable is not empty before doing
the indirect expansion.

Change-Id: Ie6e235f76809c40f36735b2e5844f668f89eff11
Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>