buid-uefi: remove the dependency on uefi-tools
Instead of using helper scripts in uefi-tools to build the edk2 targets,
add support for building edk2 targets into build-uefi script. This
allows moving away from depending on the use of uefi-tools repo.

In addition to this, update the build configs for each platform to
specify the platform description files that have to be built. While at
it, fix a typo in the array name used in the config file of rdn1edgex2
platform.

Signed-off-by: Thomas Abraham <thomas.abraham@arm.com>
Change-Id: I4d73712f90540dc7a98f97528972b6c215c1bdc3