build-scp.sh: provide product specific parameters
An experimental patch in SCP has been introduced to allow SCP build
system to accept product specific build options. This SCP patch is not
upstream yet and it is uncertain if it will get upstreamed.

For now, update the SCP firmware build script to allow the usage of the
changes introduced by this SCP experimental patch. That is, allow SCP
build script to supply additional product specific build options.

Signed-off-by: Aditya Angadi <aditya.angadi@arm.com>
Change-Id: I7b320766693147e81cc45a24f1f1b4e827c9bd55