configs/rdn2: remove console argument from grub config
Linux assigns the ttyAMA index in the order in which the uarts are
defined in the SSDT ACPI table. Also, the default uart to use is defined
in the DBG2 ACPI table and firmware will ensure that the uart used in
the DBG2 table is initialized. Passing "console=ttyAMA0,115200" as
command line parameter to linux will not honor the uart defined in the
DBG2 table. In some scenarios, ttyAMA0 might not be same uart mentioned
in the DBG2 table and might not be initialized by the boot firmware
which would cause the kernel boot to fail. To avoid this, remove
"console=ttyAMA0,115200" as command line parameter and let linux use the
uart mentioned in the DBG2 table for console.

Signed-off-by: Shriram K <shriram.k@arm.com>
Change-Id: I1f265eb30532b92ce8a54d0cc889190eb3582c9b