build-uefi.sh: print build and clean commands
Print required build and clean commands along with the time stamp. The
required commands are printed to the terminal using `set -x` which
enables a mode of the shell where all executed commands are printed to
the terminal with prefix `++`. Once the required commands are printed,
`set +x` is used to disable the prints. To avoiding printing `set +x`
itself, it is routed to /dev/null.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I2d912b5077928bf51ebb18c7f2177d9000359c86