configs/rdn2: Enable building of kvmtool and kvm-unit-tests for rdn2
Define a flag and enable it to build kvmtool and kvm-unit-tests for
rdn2 platform. Also define a variable describing the kvmtool supported
platform to help decide output directory while packaging the binary.
Add the kvmtool/kvm-unit-tests build script to the original BUILD_SCRIPTS
list as well allowing it to be the part of whole stack building.

Signed-off-by: Sathyam Panda <sathyam.panda@arm.com>
Change-Id: I9d3d3a094f440532e3cd7bad2e40f62a89f793aa