configs/rdn2cfg1: Enable building of kvmtool and kvm-unit-tests for rdn2cfg1
Define a flag and enable it to build kvmtool and kvm-unit-tests for
rdn2cfg1 platform. Also define a variable describing the kvmtool supported
platform to help decide output directory while packaging the binary.
Add the kvmtool/kvm-unit-tests build script to the original BUILD_SCRIPTS
list as well allowing it to be the part of whole stack building.

Signed-off-by: Sathyam Panda <sathyam.panda@arm.com>
Change-Id: I21c93f7d01995be369794fd41e0e0e5b5355c56a