common.base: detect host machine architecture
The machine architecture of the host machine is identified and store it
in the environment variable BUILD_MACHINE_ARCH. This variable is defined
in common.base file so it can be accessible for all platforms.

Also, CROSS_COMPILE and LINUX_COMPILER environment variables are defined
only for x86_64 host architecture.

Change-Id: Id64b26c1ab3d75dba9c5e9151bb8a6fb113312b7
Signed-off-by: Tony K Nadackal <tony.nadackal@arm.com>