build-scp.sh: build ramfw at the beginning of the build
SCP RAM firmware is where majority of the development happens, followed
by MCP RAM firmware. If there is any build error, one has to wait till
the other firmware components are built and see the error at the end of
the build. Instead, build the SCP and MCP RAM firmware at the beginning
of the build to align with the priority of the development.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I79ea44692b142af81ad5433ca582d10654ac6a12