build-test-*: use mkfs.ext3 command to create ext3 partition
Previously mkfs.ext3 was used to create an ext3 partition on a
zero-filled raw image file and then mount that partition to a mount
point using fuse-ext2 tool. The Linux and ramdisk images were then
copied to this mount point. fusermount tool was used to sync the
contents and unmount the partition. Also, tune2fs utility was needed to
assign the UUID for this partition.

fuse-ext2 tool is not available for AArch64 host machines by default. To
avoid this dependency, a single mkfs.ext3 command is used to create an
ext3 partition with UUID out of the raw image file.

Signed-off-by: Vijayenthiran Subramanaiam <vijayenthiran.subramaniam@arm.com>
Change-Id: I623923c1de21f08dabe8fe8b82659eaccf7739e0