sgi/build-test-acs.sh: refactor acs build script
This patch introduces two major changes to build-test-acs.sh.

The first change is to build only the platform binaries and not to build
the ACS disk image from source. And hence recommend to use the prebuilt
ACS image downloaded from github for the test. The prebuilt image is
available at:
https://github.com/ARM-software/arm-systemready/tree/main/SR

The second change is to disable RAS functionality, as RAS requires the
external aborts and SErrors to be handled by EL3. This is not
recommended for SBSA test.

Change-Id: Ie84cc1ec00d4968f4e45bc919de78386476a73b3
Signed-off-by: Pranav Madhu <pranav.madhu@arm.com>