build-buildroot.sh: add support to build config fragments
Buildroot can be build on both x86_64 and aarch64 host machines. The
config files required to build on either of these host machines can be
different. In order to support build of buildroot on both x86_64 and
aarch64 host machines, add support to pick config fragments specific to
the host machines and merge them with the base configuration. Once the
merge is done, delete the config file which is copied to the buildroot's
configs folder.

With this support, $BUILDROOT_DEFCONFIG file will hold the configs that
is common across both architectures. $BUILDROOT_X86_64_HOST_CONFIG file
can hold configs that are specific to x86_64 host and
$BUILDROOT_AARCH64_HOST_CONFIG file can hold configs that are specific
to aarch64 host.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: Id39a7ad3379348ae53b37309707b49f15b45987c