configs/*: add x86_64 host specific config fragment for buildroot
When building buildroot on a x86_64 host, BR2_TOOLCHAIN_EXTERNAL should
be set to y to indicate that an external toolchain should be used to
build for aarch64 target. Add a new x86_64 host specific config fragment
for buildroot and move out the x86_64 host specific config into this
file from the defconfig file.

Signed-off-by: Vijayenthiran Subramaniam <vijayenthiran.subramaniam@arm.com>
Change-Id: I379414f48d37884834a214400e933aba5b1e5af5