build-grub.sh: build grub with python3
GNU GRUB autogen.sh defaults to using the python command which defaults
to python2 version. This is not available on newer Linux distributions
since the EOL of python2 [1]. To resolve this, override the PYTHON
environment variable to python3 when building grub.

[1]: https://www.python.org/doc/sunset-python-2/

Note: this change will not be required when grub is updated with the
latest version which includes following commit:
https://git.savannah.gnu.org/cgit/grub.git/commit/?id=8f35208db4154ede690a18e408375815bbf320b7

Signed-off-by: Joel Goddard <Joel.Goddard@arm.com>
Change-Id: I8b9b36e5ba97e07b50ecb5bdba1ac3ce90be043b