configs/rdv2: Enable building of kvmtool and kvm-unit-tests for rdv2
Define a flag and enable it to build kvmtool and kvm-unit-tests for
rdv2 platform. Also define a variable describing the kvmtool supported
platform to help decide output directory while packaging the binary.
Add the kvmtool/kvm-unit-tests build script to the original BUILD_SCRIPTS
list as well allowing it to be the part of whole stack building.

Signed-off-by: Sathyam Panda <sathyam.panda@arm.com>
Change-Id: I55cf58c31fa038d0fefcbf1697f839296a29ae81